+++ src/Money.php
@@ -1,42 +1,26 @@
 <?php
+
 namespace Ayeo\Price;
 
 class Money
 {
-	/**
-	 * @var float
-	 */
-	private $value;
+	private float $value;
 
-	/**
-	 * @param float $value
-	 */
-	public function __construct($value)
+	public function __construct(float $value)
 	{
-		if (is_numeric($value) === false) {
-			throw new \LogicException('Money value must be numeric');
-		}
-
-		if ($value < 0) {
+		if ($value < 0.) {
 			throw new \LogicException('Money value must be positive');
 		}
 
-		$this->value = round((float) $value, 2);
+		$this->value = $value;
 	}
 
-	/**
-	 * @return float
-	 */
-	public function getValue()
+	public function getValue(): float
 	{
 		return $this->value;
 	}
 
-	/**
-	 * @param Money $money
-	 * @return bool
-	 */
-	public function isGreaterThan(Money $money)
+	public function isGreaterThan(Money $money): bool
 	{
 		//floating point calculations precision problem here
 		return round($this->getValue(), 6) > round($money->getValue(), 6);