+++ tests/PriceTest.php
@@ -1,16 +1,16 @@
 <?php
+
 namespace Ayeo\Price\Test;
 
 use Ayeo\Price\Price;
 use LogicException;
+use PHPUnit\Framework\TestCase;
 
-class PriceTest extends \PHPUnit_Framework_TestCase
+class PriceTest extends TestCase
 {
-    /**
-     * @expectedException \LogicException
-     */
     public function testBuildInvalidPrice()
     {
+        $this->expectException(LogicException::class);
         new Price(120.00, 100.00, "PLN", 23);
     }
 
@@ -23,7 +23,7 @@ class PriceTest extends \PHPUnit_Framework_TestCase
 
         $this->assertEquals(300.00, $result->getNett());
         $this->assertEquals(360.00, $result->getGross());
-        $this->assertEquals(20, $result->getTaxValue());
+        $this->assertEquals(20, $result->getTaxRate());
         //$this->assertEquals(true, $result->hasTaxRate());
     }
 
@@ -69,7 +69,7 @@ class PriceTest extends \PHPUnit_Framework_TestCase
 	public function testCreating($nett, $gross, $tax)
 	{
 		$price = new Price($nett, $gross, 'USD', $tax);
-        $this->assertEquals($tax, $price->getTaxValue());
+        $this->assertEquals($tax, $price->getTaxRate());
 	}
 
     /**
@@ -133,11 +133,11 @@ class PriceTest extends \PHPUnit_Framework_TestCase
 		$this->assertEquals($expectedGross, $A->add($B)->getGross());
 		$this->assertEquals($expectedGross, $B->add($A)->getGross());
 
-		$this->assertEquals($tax, $B->add($A)->getTaxValue());
-		$this->assertEquals($tax, $A->add($B)->getTaxValue());
+		$this->assertEquals($tax, $B->add($A)->getTaxRate());
+		$this->assertEquals($tax, $A->add($B)->getTaxRate());
 
-		$this->assertEquals($nettA, $A->getNett(), '', 0.01);
-		$this->assertEquals($nettB, $B->getNett(), '', 0.01);
+		$this->assertEqualsWithDelta($nettA, $A->getNett(), 0.01);
+		$this->assertEqualsWithDelta($nettB, $B->getNett(), 0.01);
 	}
 
 	public function addingDataProvider()
@@ -172,38 +172,29 @@ class PriceTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('USD', $C->getCurrencySymbol());
     }
 
-    /**
-     * @expectedException           LogicException
-     * @expectedExceptionMessage    Can not operate on different currencies ("USD" and "GBP")
-     */
     public function testAddingDifferentCurrencies()
     {
+        $this->expectExceptionMessage('Can not operate on different currencies ("USD" and "GBP")');
         $A = new Price(100, 130, 'USD');
         $B = new Price(300, 330, 'GBP');
         $A->add($B);
     }
 
-    /**
-     * @expectedException           LogicException
-     * @expectedExceptionMessage    Nett must not be greater than gross
-     */
     public function testNettGreaterThanGross()
     {
+        $this->expectExceptionMessage('Nett must not be greater than gross');
         new Price(100.00, 90.00, 'USD');
     }
 
     public function testNettSameAsGross()
     {
         $price = new Price(100.00, 100.00, 'USD', 0);
-        $this->assertEquals(0, $price->getTaxValue());
+        $this->assertEquals(0, $price->getTaxRate());
     }
 
-    /**
-     * @expectedException           LogicException
-     * @expectedExceptionMessage    Invalid currency symbol: "PLNG"
-     */
     public function testInvalidCurrencySymbol()
     {
+        $this->expectExceptionMessage('Invalid currency symbol: "PLNG"');
         new Price(100, 200, 'PLNG');
     }
 
@@ -258,39 +249,30 @@ class PriceTest extends \PHPUnit_Framework_TestCase
         ];
     }
 
-    /**
-     * @expectedException           LogicException
-     * @expectedExceptionMessage    Money value must be positive
-     */
     public function testNegativeNett()
     {
+        $this->expectExceptionMessage('Money value must be positive');
         new Price(-10.00, 20, 'USD');
     }
 
-    /**
-     * @expectedException           LogicException
-     * @expectedExceptionMessage    Money value must be positive
-     */
     public function testNegativeNettAndGross()
     {
+        $this->expectExceptionMessage('Money value must be positive');
         new Price(10.00, -15.00, 'USD');
     }
 
     public function  testSubtractNett()
     {
-        $price = new Price(13.34, 15.53, 'USD');
-        $result = $price->subtractNett(10.00, 'USD');
+        $price = new Price(10, 12, 'USD');
+        $result = $price->subtractNett(5.00, 'USD');
 
-        $this->assertEquals(3.34, $result->getNett());
-        $this->assertEquals(3.87, $result->getGross());
+        $this->assertEquals(5, $result->getNett());
+        $this->assertEquals(6, $result->getGross());
     }
 
-    /**
-     * @expectedException           LogicException
-     * @expectedExceptionMessage    Money value must be positive
-     */
     public function testSubtractNegativeNett()
     {
+        $this->expectExceptionMessage('Money value must be positive');
         $price = new Price(13.34, 15.53, 'USD');
         $price->subtractNett(-10.00, 'USD');
     }
@@ -305,16 +287,6 @@ class PriceTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue($price->isEqual($newPrice));
     }
 
-    /**
-     * @expectedException           LogicException
-     * @expectedExceptionMessage    Money value must be numeric
-     */
-    public function testSubtractStringNett()
-    {
-        $price = new Price(13.34, 15.53, 'USD');
-        $price->subtractNett("number", 'USD');
-    }
-
     public function  testSubtractGross()
     {
         $price = new Price(13.34, 15.53, 'USD');
@@ -323,12 +295,9 @@ class PriceTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(5.53, $result->getGross());
     }
 
-    /**
-     * @expectedException           LogicException
-     * @expectedExceptionMessage    Money value must be positive
-     */
     public function testSubtractNegativeGross()
     {
+        $this->expectExceptionMessage('Money value must be positive');
         $price = new Price(13.34, 15.53, 'USD');
         $price->subtractGross(-10.00, 'USD');
     }
@@ -342,17 +311,6 @@ class PriceTest extends \PHPUnit_Framework_TestCase
         $newPrice = $price->subtractGross(0.00, 'USD');
         $this->assertTrue($price->isEqual($newPrice));
     }
-
-    /**
-     * @expectedException           LogicException
-     * @expectedExceptionMessage    Money value must be numeric
-     */
-    public function testSubtractString()
-    {
-        $price = new Price(13.34, 15.53, 'USD');
-        $price->subtractGross("number", 'USD');
-    }
-
     public function testAddGross()
     {
         $A = new Price(13.24, 20.99, 'USD');
@@ -370,36 +328,15 @@ class PriceTest extends \PHPUnit_Framework_TestCase
 
         $this->assertEquals(246.00, $result->getGross());
         $this->assertEquals(200.00, $result->getNett());
-        $this->assertEquals(23, $result->getTaxValue());
+        $this->assertEquals(23, $result->getTaxRate());
         $this->assertEquals('USD', $result->getCurrencySymbol());
 
     }
 
-    /**
-     * @expectedException           LogicException
-     * @expectedExceptionMessage    Tax percent must positive
-     */
     public function testBuildByNettUsingNegativeTax()
     {
+        $this->expectExceptionMessage('Tax percent must positive');
         Price::buildByNett(100.00, -2);
-    }
-
-    /**
-     * @expectedException           LogicException
-     * @expectedExceptionMessage    Tax percent must be integer
-     */
-    public function testBuildByNettUsingNonIntegerLikeTax()
-    {
-        Price::buildByNett(100.00, 2.02);
-    }
-
-    /**
-     * @expectedException           LogicException
-     * @expectedExceptionMessage    Tax percent must be integer
-     */
-    public function testBuildByNettUsingNonIntegerLikeStringTax()
-    {
-        Price::buildByNett(100.00, "2.02", 'USD');
     }
 
     public function testBuildByNettUsingIntegerTax()
@@ -468,12 +405,9 @@ class PriceTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(23, $price->getTaxRate());
     }
 
-    /**
-     * @expectedException           \LogicException
-     * @expectedExceptionMessage    Can not operate on different currencies ("USD" and "GBP")
-     */
     public function testAddDifferentCurrencies()
     {
+        $this->expectExceptionMessage('Can not operate on different currencies ("USD" and "GBP")');
         $usd = Price::buildByGross(100.00, 8, 'USD');
         $eur = Price::buildByGross(100.00, 8, 'GBP');
         $usd->add($eur);
@@ -486,27 +420,17 @@ class PriceTest extends \PHPUnit_Framework_TestCase
 //        $price = new Price(100.00, 120.00, "USD", 10);
 //    }
 
-	/**
-	 * @expectedException LogicException
-	 */
-    public function testFloatTaxRate()
-    {
-        $price = new Price(100.00, 120.00, "USD", 19.99);
-    }
-
     public function testFluentInterface()
     {
     	$price = Price::buildByGross(100, 23, "PLN");
     	$newPrice = $price->add($price)->multiply(2)->divide(3);
 
-    	$this->assertEquals((100+100)*2/3, $newPrice->getGross(), "", 0.01);
+    	$this->assertEqualsWithDelta((100+100)*2/3, $newPrice->getGross(),  0.01);
     }
 
-	/**
-	 * @expectedException LogicException
-	 */
     public function testCreateEmptyPriceWithNoCurrency()
     {
+        $this->expectException(LogicException::class);
     	$price = Price::buildEmpty();
     	$price->getCurrency();
     }
@@ -554,13 +478,10 @@ class PriceTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(0, $results->getGross());
         $this->assertEquals(23, $results->getTaxRate());
     }
-    
-    /**
-     * @expectedException LogicException
-     * @expectedExceptionMessage Multiply param must greater than 0
-     */
+
     public function testMultipleByNegativeValue()
     {
+        $this->expectExceptionMessage('Multiply param must greater than 0');
         $price = Price::buildByNett(10, 23, 'PLN');
         $price->multiply(-5);
     }